Move AH intrp stuff out of ParallelInterpolation
