+++ src/ApparentHorizons/Callbacks/ErrorOnFailedApparentHorizon.hpp
