+++ src/ApparentHorizons/Callbacks/IgnoreFailedApparentHorizon.hpp
