+++ src/ParallelAlgorithms/Interpolation/Targets/CMakeLists.txt
@@ -5,7 +5,6 @@ spectre_target_sources(
   ${LIBRARY}
   PRIVATE
   AngularOrdering.cpp
-  ApparentHorizon.cpp
   KerrHorizon.cpp
   LineSegment.cpp
   SpecifiedPoints.cpp
@@ -18,7 +17,6 @@ spectre_target_headers(
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
   AngularOrdering.hpp
-  ApparentHorizon.hpp
   KerrHorizon.hpp
   LineSegment.hpp
   SpecifiedPoints.hpp
